feat(service-worker): add function to provide service worker
add function `provideServiceWorker` to register a service worker in standalone applications without using `ServiceWorkerModule.register()`
