+++ packages/service-worker/src/index.ts
@@ -6,15 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
 export {NoNewVersionDetectedEvent, UnrecoverableStateEvent, UpdateActivatedEvent, UpdateAvailableEvent, VersionDetectedEvent, VersionEvent, VersionInstallationFailedEvent, VersionReadyEvent} from './low_level';
-export {ServiceWorkerModule, SwRegistrationOptions} from './module';
+export {ServiceWorkerModule} from './module';
+export {provideServiceWorker, SwRegistrationOptions} from './provider';
 export {SwPush} from './push';
 export {SwUpdate} from './update';
